feat(dashboard): React hook form with Zod

+++ src/actions/actions.ts
@@ -4,15 +4,32 @@ import prisma from "@/lib/db";
 import { revalidatePath } from "next/cache";
 import { PetEssentials } from "@/lib/types";
 import { Pet } from "@prisma/client";
+import { petFormSchema, petIdSchema } from "@/lib/validation";
+import { getPetById } from "@/lib/server-utils";
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (!validatedPet.success) {
+    return {
+      message: "Invalid pet data.",
+    };
+  }
+
   try {
     await prisma.pet.create({
-      data: pet,
+      data: {
+        ...validatedPet.data,
+        // user: {
+        //   connect: {
+        //     id: session.user.id,
+        //   },
+        // },
+      },
     });
-  } catch (err) {
+  } catch (error) {
+    console.log(error);
     return {
-      message: "Error adding pet",
+      message: "Could not add pet.",
     };
   }
 
@@ -20,32 +37,64 @@ export async function addPet(pet: PetEssentials) {
 }
 
 export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  const validatedPet = petFormSchema.safeParse(newPetData);
+
+  if (!validatedPetId.success || !validatedPet.success) {
+    return {
+      message: "Invalid pet data.",
+    };
+  }
+
+  // Check if pet exists:
+  const pet = await getPetById(validatedPetId.data);
+  if (!pet) {
+    return {
+      message: "Pet not found.",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
-  } catch (err) {
+  } catch (error) {
     return {
-      message: "Error editing pet",
+      message: "Could not edit pet.",
     };
   }
 
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data.",
+    };
+  }
+
+  // Check if pet exists:
+  const pet = await getPetById(validatedPetId.data);
+  if (!pet) {
+    return {
+      message: "Pet not found.",
+    };
+  }
+
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
-  } catch (err) {
+  } catch (error) {
     return {
-      message: "Error deleting a pet",
+      message: "Could not delete pet.",
     };
   }
